// marketplace/src/types/custom-interface.ts
export interface ISignUpFormData {
  username: string;
  email: string;
  password: string;
  confirmPassword: string;
  remember?: boolean;
}